requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nodejs-wheel-binaries==22.20.0
numba==0.62.1
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
pandas==2.3.3
propcache==0.4.1
python-dateutil==2.9.0.post0
pytubefix==10.3.3
pytz==2025.2
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
